All indices from int to size_t

+++ HungarianAlg/HungarianAlg.h
@@ -13,22 +13,22 @@ private:
 	// --------------------------------------------------------------------------
 	// Computes the optimal assignment (minimum overall costs) using Munkres algorithm.
 	// --------------------------------------------------------------------------
-	void assignmentoptimal(assignments_t& assignment, track_t *cost, track_t *distMatrix, int nOfRows, int nOfColumns);
-	void buildassignmentvector(assignments_t& assignment, bool *starMatrix, int nOfRows, int nOfColumns);
-	void computeassignmentcost(const assignments_t& assignment, track_t *cost, track_t *distMatrix, int nOfRows);
-	void step2a(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim);
-	void step2b(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim);
-	void step3(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim);
-	void step4(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim, int row, int col);
-	void step5(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, int nOfRows, int nOfColumns, int minDim);
+	void assignmentoptimal(assignments_t& assignment, track_t *cost, track_t *distMatrix, size_t nOfRows, size_t nOfColumns);
+	void buildassignmentvector(assignments_t& assignment, bool *starMatrix, size_t nOfRows, size_t nOfColumns);
+	void computeassignmentcost(const assignments_t& assignment, track_t *cost, track_t *distMatrix, size_t nOfRows);
+	void step2a(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim);
+	void step2b(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim);
+	void step3(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim);
+	void step4(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim, size_t row, size_t col);
+	void step5(assignments_t& assignment, track_t *distMatrix, bool *starMatrix, bool *newStarMatrix, bool *primeMatrix, bool *coveredColumns, bool *coveredRows, size_t nOfRows, size_t nOfColumns, size_t minDim);
 	// --------------------------------------------------------------------------
 	// Computes a suboptimal solution. Good for cases with many forbidden assignments.
 	// --------------------------------------------------------------------------
-	void assignmentsuboptimal1(assignments_t& assignment, track_t *cost, track_t *distMatrixIn, int nOfRows, int nOfColumns);
+	void assignmentsuboptimal1(assignments_t& assignment, track_t *cost, track_t *distMatrixIn, size_t nOfRows, size_t nOfColumns);
 	// --------------------------------------------------------------------------
 	// Computes a suboptimal solution. Good for cases with many forbidden assignments.
 	// --------------------------------------------------------------------------
-	void assignmentsuboptimal2(assignments_t& assignment, track_t *cost, track_t *distMatrixIn, int nOfRows, int nOfColumns);
+	void assignmentsuboptimal2(assignments_t& assignment, track_t *cost, track_t *distMatrixIn, size_t nOfRows, size_t nOfColumns);
 
 public:
 	enum TMethod
